Finish moving debug utils from foundation to core
