+++ source/library/concurrency/private/thread_util.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2016,2019 Sean Middleditch, all rights reserverd.
 
 #include "potato/concurrency/thread_util.h"
-#include "potato/foundation/assertion.h"
+#include <potato/core/assertion.h>
 #include <atomic>
 #include <limits>
