+++ source/library/core/private/debug.posix.cpp
@@ -1,6 +1,5 @@
 // Copyright (C) 2014 Sean Middleditch, all rights reserverd.
 
-#include "potato/foundation/string_format.h"
 #include <cstdio>
 #include <cstdlib>
