+++ source/library/core/public/potato/core/assertion.h
@@ -2,10 +2,9 @@
 
 #pragma once
 
-#include "potato/core/debug.h"
-#include "platform.h"
-#include "string_format.h"
-#include "fixed_string_writer.h"
+#include <potato/core/debug.h>
+#include <potato/foundation/string_format.h>
+#include <potato/foundation/fixed_string_writer.h>
 #include <utility>
 
 #if defined(NDEBUG)
