+++ source/library/foundation/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.11)
 project(foundation VERSION 0.1 LANGUAGES CXX)
 
-add_library(foundation)
+add_library(foundation STATIC)
 add_library(potato::foundation ALIAS foundation)
 
 include(up_set_common_properties)
@@ -10,7 +10,7 @@ up_set_common_properties(foundation)
 target_link_libraries(foundation
     PUBLIC
         formatxx
-    PRIVATE
+    INTERFACE
         core
 )
