--- source/library/foundation/private/CMakeLists.txt
@@ -1,3 +1,0 @@
-target_sources(foundation PRIVATE
-    ${CMAKE_CURRENT_SOURCE_DIR}/string_writer.cpp
-)
