+++ source/library/gpu/private/d3d11_backend/d3d11_platform.cpp
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Sean Middleditch, all rights reserverd.
 
 #include "d3d11_platform.h"
-#include "potato/foundation/assertion.h"
+#include <potato/core/assertion.h>
 
 auto up::d3d11::toNative(GpuShaderSemantic semantic) noexcept -> zstring_view {
     switch (semantic) {
